backend/requirements.txt
fastapi
uvicorn
playwright
axe-playwright-python
google-generativeai
pydantic
pymongo[srv]
tenacity==9.1.4